+++ themes/default/templates/core/blocks/theme_menu_block.tpl
@@ -18,7 +18,7 @@
                     <a href="index.php?page=customer:search"><img src="{$theme_images_dir}icons/16x16/viewmag.gif" alt="" border="0" height="14" width="14" /> {t}Search Customers{/t}</a>
                     {if $customer_id > 0 }
                         <a href="index.php?page=customer:edit&customer_id={$customer_id}"><img src="{$theme_images_dir}icons/edit_employee.gif" alt="" border="0" height="14" width="14" /> {t}Edit Customer{/t}</a>                        
-                        <a href="index.php?page=customer:delete&customer_id={$customer_id}"><img src="{$theme_images_dir}icons/delete_employees.gif" alt="" border="0" height="14" width="14" /> {t}Delete Customer{/t}</a>
+                        <a href="index.php?page=customer:delete&customer_id={$customer_id}" onclick="return confirmDelete('{t}Are you sure you want to delete this customer?{/t}');"><img src="{$theme_images_dir}icons/delete_employees.gif" alt="" border="0" height="14" width="14" /> {t}Delete Customer{/t}</a>
                         <a href="index.php?page=customer:email&customer_id={$customer_id}"><img src="{$theme_images_dir}icons/16x16/email.jpg" alt="" border="0" height="14" width="14" /> {t}Email Customer{/t}</a>
                         <a href="index.php?page=user:new&customer_id={$customer_id}"><img src="{$theme_images_dir}icons/16x16/email.jpg" alt="" border="0" height="14" width="14" /> {t}New Customer Login{/t}</a>
                         <a href="index.php?page=customer:note_new&customer_id={$customer_id}"><img src="{$theme_images_dir}icons/16x16/email.jpg" alt="" border="0" height="14" width="14" />{t}New Note{/t}</a>
@@ -26,7 +26,7 @@
                     {/if}
                         <a href="index.php?page=giftcert:search"><img src="{$theme_images_dir}icons/gift.png" alt="" border="0" height="14" width="14" /> {t}Search Gift Certificates{/t}</a>
                     {if $giftcert_id > 0}
-                        <a href="index.php?page=giftcert:edit&giftcert_id={$giftcert_id}"><img src="{$theme_images_dir}icons/gift.png" alt="" border="0" height="14" width="14" /> {t}Search Gift Certificates{/t}</a>
+                        <a href="index.php?page=giftcert:edit&giftcert_id={$giftcert_id}"><img src="{$theme_images_dir}icons/gift.png" alt="" border="0" height="14" width="14" /> {t}Edit Gift Certificates{/t}</a>
                     {/if}
                 </div>
 
@@ -72,7 +72,7 @@
                         {if $expense_id > 0 }
                             <a href="index.php?page=expense:details&expense_id={$expense_id}"><img src="{$theme_images_dir}icons/view.gif" alt="" border="0" height="14" width="14" /> {t}Expense Details{/t}</a>
                             <a href="index.php?page=expense:edit&expense_id={$expense_id}"><img src="{$theme_images_dir}icons/edit.gif" alt="" border="0" height="14" width="14" /> {t}Edit Expense{/t}</a>
-                            <a href="index.php?page=expense:delete&expense_di={$expense_id}" onclick="return confirmDelete('{t}Are you Sure you want to delete this Expense Record? This will permanently remove the record from the database.{/t}');"><img src="{$theme_images_dir}icons/delete.gif" alt="" border="0" height="14" width="14" /> {t}Delete Expense{/t}</a>
+                            <a href="index.php?page=expense:delete&expense_di={$expense_id}" onclick="return confirmDelete('{t}Are you sure you want to delete this Expense Record?{/t}');"><img src="{$theme_images_dir}icons/delete.gif" alt="" border="0" height="14" width="14" /> {t}Delete Expense{/t}</a>
                         {/if}
 
                         <!-- Refunds -->
@@ -81,7 +81,7 @@
                         {if $refund_id > 0 }
                             <a href="index.php?page=refund:details&refund_id={$refund_id}"><img src="{$theme_images_dir}icons/view.gif" alt="" border="0" height="14" width="14" /> {t}Refund Details{/t}</a>
                             <a href="index.php?page=refund:edit&refund_id={$refund_id}"><img src="{$theme_images_dir}icons/edit.gif" alt="" border="0" height="14" width="14" /> {t}Edit Refund{/t}</a>
-                            <a href="index.php?page=refund:delete&refund_id={$refund_id}" onclick="return confirmDelete('{t}Are you Sure you want to delete this Refund Record? This will permanently remove the record from the database.{/t}');"><img src="{$theme_images_dir}icons/delete.gif" alt="" border="0" height="14" width="14" /> {t}Delete Refund{/t}</a>
+                            <a href="index.php?page=refund:delete&refund_id={$refund_id}" onclick="return confirmDelete('{t}Are you sure you want to delete this Refund Record?{/t}');"><img src="{$theme_images_dir}icons/delete.gif" alt="" border="0" height="14" width="14" /> {t}Delete Refund{/t}</a>
                         {/if}
                         <a href="index.php?page=report:financial"><img src="{$theme_images_dir}icons/reports.png" alt="" border="0" height="14" width="14" /> {t}Financial Report{/t}</a>
                         
@@ -98,7 +98,7 @@
                         {if $supplier_id > 0 }
                             <a href="index.php?page=supplier:details&supplier_id={$supplier_id}"><img src="{$theme_images_dir}icons/view.gif" alt="" border="0" height="14" width="14" /> {t}Supplier Details{/t}</a>
                             <a href="index.php?page=supplier:edit&supplier_id={$supplier_id}"><img src="{$theme_images_dir}icons/edit.gif" alt="" border="0" height="14" width="14" /> {t}Edit Supplier{/t}</a>
-                            <a href="index.php?page=supplier:delete&supplier_id={$supplier_id}" onclick="return confirmDelete('{t}Are you Sure you want to delete this Supplier Record? This will permanently remove the record from the database.{/t}');"><img src="{$theme_images_dir}icons/delete.gif" alt="" border="0" height="14" width="14" /> {t}Delete Supplier{/t}</a>
+                            <a href="index.php?page=supplier:delete&supplier_id={$supplier_id}" onclick="return confirmDelete('{t}Are you Sure you want to delete this Supplier?{/t}');"><img src="{$theme_images_dir}icons/delete.gif" alt="" border="0" height="14" width="14" /> {t}Delete Supplier{/t}</a>
                         {/if}
                     </div>
                 {/if}
@@ -126,7 +126,7 @@
                         <a href="index.php?page=user:new"><img src="{$theme_images_dir}icons/16x16/view.gif" alt="" border="0" height="14" width="14" /> {t}New Employee{/t}</a>
                         {if $user_id > 0 }
                             <a href="index.php?page=user:edit&user_id={$user_id}"><img src="{$theme_images_dir}icons/edit_employee.gif" alt="" border="0" height="14" width="14" /> {t}Edit User{/t}</a>
-                            <a href="index.php?page=user:delete&user_id={$user_id}"><img src="{$theme_images_dir}icons/delete_employees.gif" alt="" border="0" height="14" width="14" /> {t}Delete User{/t}</a>
+                            <a href="index.php?page=user:delete&user_id={$user_id}" onclick="return confirmDelete('{t}Are you sure you want to delete this user?{/t}');"><img src="{$theme_images_dir}icons/delete_employees.gif" alt="" border="0" height="14" width="14" /> {t}Delete User{/t}</a>
                         {/if}
                         <a href="index.php?page=administrator:acl"><img src="{$theme_images_dir}icons/encrypted.png" alt="" border="0" height="14" width="14" /> {t}Permissions{/t}</a>
